requirements.txt
fastapi
uvicorn
pymongo
google-generativeai
mistralai
PyPDF2
python-dotenv